requirements.txt
numpy==1.26.4
openai==1.37.0
tiktoken==0.7.0
streamlit==1.36.0
streamlit-feedback==0.1.3
bcrypt==4.2.0
pymongo==4.6.1
protobuf==4.25.1
captcha==0.6.0
weaviate-client==4.5.0
ijson==3.3.0
sentence-transformers==3.0.1
nest-asyncio==1.6.0
--pre bibtexparser
latexcodec==3.0.0
pylatexenc==3.0a25
flask==3.0.3
python-dotenv==1.0.1
torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu